+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.4.1](../../tree/v0.4.1) - 30.07.2020
+* :tada: **New** option `allowDuplicates` now functions as expected
+* :tada: **New** option `capitalize` now functions as expected
+* :tada: **New** option `placeholderNotation`
+* :boom: **Changed** option `preserveCurlyBrackets` to `preservePlaceholderNotation`
+* :bug: **Fix** `setOptions` so that it no longer resets unspecified options to previous defaults
+
 ## [0.4.0](../../tree/v0.4.0) - 26.07.2020
 #### Features
 * :tada: **New** option `forceNewSentence`, which will force `Sentence.generate()` to avoid duplicate of previous sentence when possible